% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_by_unicode.R
\name{write_by_unicode}
\alias{write_by_unicode}
\title{write by unicode}
\usage{
write_by_unicode(char)
}
\arguments{
\item{char}{char vector}
}
\value{
non print string into stdout
}
\description{
to chage non ascii character into unicode
}
